Rename client model
