+++ app/models/client.js
@@ -1,19 +1,18 @@
-var mongoose = require('mongoose');
-var crypto = require('crypto');
-var Schema = mongoose.Schema;
+var mongoose = require('mongoose')
+var Schema = mongoose.Schema
 
 
-var apiClientSchema = new Schema({
+var clientSchema = new Schema({
   name: { type: String, default: '' },
   user: { type: Schema.ObjectId, ref: 'User' },
   secret: { type: String, default: '' },
   redirectURI: { type: String, default: '' },
   createdAt: { type: Date, default: Date.now }
-});
+})
 
-var APIClient = module.exports = mongoose.model('APIClient', apiClientSchema);
+var Client = module.exports = mongoose.model('Client', clientSchema)
 
-apiClientSchema.set('toJSON', {
+clientSchema.set('toJSON', {
   versionKey: false,
   transform: function (doc, ret) {
     delete ret.secret
